Use long instead of int for monetary values in Bank, CashMachine, and Wallet
Prevents integer overflow on large transactions. Added null-checks in
Cashier constructor and withdraw method.

+++ kodilla-basic-tests/src/main/java/com/kodilla/bank/Bank.java
@@ -14,8 +14,8 @@ public class Bank {
         cashMachines.add(cashMachine);
     }
 
-    public int getBalance() {
-        int sum = 0;
+    public long getBalance() {
+        long sum = 0;
         for (CashMachine cashMachine : cashMachines) {
             sum += cashMachine.getBalance();
         }
