+++ kodilla-cucumber/src/main/java/wallet/InsufficientFundsException.java
@@ -2,7 +2,8 @@ package wallet;
 
 public class InsufficientFundsException extends RuntimeException {
 
-    public InsufficientFundsException(int requested, int available) {
-        super("Insufficient funds: requested " + requested + ", available " + available);
+    public InsufficientFundsException(long requested, long available) {
+        super("Insufficient funds: requested " + requested
+                + ", available " + available);
     }
 }
