+++ kodilla-cucumber/src/test/java/wallet/WalletSteps.java
@@ -10,7 +10,7 @@ public class WalletSteps {
     private Wallet wallet;
     private CashSlot cashSlot;
     private String errorMessage;
-    private int checkedBalance;
+    private long checkedBalance;
 
     private void initIfNeeded() {
         if (wallet == null) {
@@ -23,7 +23,8 @@ public class WalletSteps {
     public void i_have_deposited_in_my_wallet(int amount) {
         initIfNeeded();
         wallet.deposit(amount);
-        assertEquals(amount, wallet.getBalance(), "Incorrect wallet balance after deposit");
+        assertEquals(amount, wallet.getBalance(),
+                "Incorrect wallet balance after deposit");
     }
 
     @Given("there is ${int} in my wallet")
@@ -63,7 +64,8 @@ public class WalletSteps {
 
     @Then("the balance of my wallet should be ${int}")
     public void the_balance_of_my_wallet_should_be(int expected) {
-        assertEquals(expected, wallet.getBalance(), "Incorrect wallet balance");
+        assertEquals(expected, wallet.getBalance(),
+                "Incorrect wallet balance");
     }
 
     @Then("nothing should be dispensed")
@@ -73,8 +75,10 @@ public class WalletSteps {
 
     @Then("I should be told that I don't have enough money in my wallet")
     public void i_should_be_told_insufficient_funds() {
-        assertNotNull(errorMessage, "Expected an error message but got none");
-        assertTrue(errorMessage.contains("Insufficient funds"), "Expected 'Insufficient funds' message");
+        assertNotNull(errorMessage,
+                "Expected an error message but got none");
+        assertTrue(errorMessage.contains("Insufficient funds"),
+                "Expected 'Insufficient funds' message");
     }
 
     @Then("I should see that the balance is ${int}")
